Use URI for uploaded files and delete request for not referenced files
The URI includes the hierarchy of uploaded files, especially for unpacked ZIP files.

+++ frontend/src/app/formly/types/table/upload-files-dialog/upload-files-dialog.component.ts
@@ -1,5 +1,5 @@
-import { Component, OnInit } from "@angular/core";
-import { MatDialogRef } from "@angular/material/dialog";
+import { Component, Inject, OnInit } from "@angular/core";
+import { MAT_DIALOG_DATA, MatDialogRef } from "@angular/material/dialog";
 import { FormStateService } from "../../../../+form/form-state.service";
 import { TransfersWithErrorInfo } from "../../../../shared/upload/TransferWithErrors";
 import { UploadService } from "../../../../shared/upload/upload.service";
@@ -9,6 +9,7 @@ import { map } from "rxjs/operators";
 
 export interface LinkInfo {
   file: string;
+  uri: string;
 }
 
 @UntilDestroy()
@@ -18,7 +19,6 @@ export interface LinkInfo {
   styleUrls: ["./upload-files-dialog.component.scss"],
 })
 export class UploadFilesDialogComponent implements OnInit {
-  droppedFiles: any = [];
   chosenFiles: TransfersWithErrorInfo[] = [];
   targetUrl = "/api/upload/";
   docId = null;
@@ -27,7 +27,8 @@ export class UploadFilesDialogComponent implements OnInit {
   constructor(
     private dlgRef: MatDialogRef<UploadFilesDialogComponent, LinkInfo[]>,
     formStateService: FormStateService,
-    private uploadService: UploadService
+    private uploadService: UploadService,
+    @Inject(MAT_DIALOG_DATA) public data: { currentItems: any[] }
   ) {
     this.docId = formStateService.getForm().get("_id").value;
     this.targetUrl += this.docId;
@@ -36,7 +37,10 @@ export class UploadFilesDialogComponent implements OnInit {
   ngOnInit(): void {}
 
   removeUploadedFile(fileId: string) {
-    this.uploadService.deleteUploadedFile(this.docId, fileId);
+    const fileNotReferenced = this.fileExistsInTable(fileId);
+    if (!fileNotReferenced) {
+      this.uploadService.deleteUploadedFile(this.docId, fileId);
+    }
   }
 
   submit() {
@@ -50,7 +54,7 @@ export class UploadFilesDialogComponent implements OnInit {
   private getSuccessfulUploadedFiles(): LinkInfo[] {
     return this.chosenFiles
       .filter((file) => file.transfer.success)
-      .map((file) => ({ file: file.transfer.name }));
+      .map((file) => ({ file: file.transfer.name, uri: file.transfer.name }));
   }
 
   private extractAndCloseDialog() {
@@ -83,12 +87,19 @@ export class UploadFilesDialogComponent implements OnInit {
   ): LinkInfo[] {
     return UploadFilesDialogComponent.flatten(
       response.map((zipFile) =>
-        zipFile.files.map((file) => ({ file: file.file }))
+        zipFile.files.map((file) => ({ file: file.file, uri: file.uri }))
       )
     );
   }
 
   private static flatten<T>(arr: T[][]): T[] {
     return ([] as T[]).concat(...arr);
+  }
+
+  private fileExistsInTable(fileId: string): boolean {
+    return (
+      this.data.currentItems.find((item) => item.link.value === fileId) !==
+      undefined
+    );
   }
 }
